Add Core API support for new constraints
This only adds supports in the read path as intended.

Align new constraints name in Core API with Cypher names
Also Align toString to Cypher representation
To make it clear in the stringify version of the constraint, what is the
entity token that defines the constraint, Cypher uses `[:%s => ]`
(notice the arrow, `=>` with nothing else on the right side).

Migrate FormatFamilyMigrationIT to Enterprise and add new constraints