Fix name generation test to use entity and token IDs
Before `assertName(...)` was ignoring the entity token and property
key tokens from the schema and passing a hardcoded String array to the
`SchemaNameUtil.generateName(...)` method.

This logic was bad, because the test would conclude that two
constraints of the same type, but targetting different entities would
actually produce the same generated name which is false.

Now we resolve/lookup the token names from the schema using the test
data and pass the complete info to the name generation logic.