5 enveloped raft log ()
Implements a new Raft Log based on EnvelopedChannels.

The new log can be enabled by setting internal.dbms.cluster.raft.log_implementation=enveloped

This is not a production ready implementation as it does not have any optimised caches and does not handle concurrency at the moment.

There are also a bunch of TODOs merged that need to be resolved.

However this implementation is functionally complete and should be able to pass any cluster IT (except for a handful of known tests)

This is mostly just new code since it is an alternative implementation to our current log. It does use the LogHeader from kernels current log, and has therefor added a term (long) to the header. It also has tidied up how we report metrics in a few places