Issue with non-existing type interpreted as any type
For dense nodes in block format a -1 sent in a RelationshipSelection was interpreted
as "any type" rather than "no type" leading to the wrong results for dense nodes in
some cases. Maybe this should also be fixed in BlockRelationshipTraversalCursor but for
future safety we should never pass in -1 to the storage layer.