Use token names instead of id for hashing the additional metadata
This changes the SchemaNameUtil.generateName() in order to use the token
names - instead of their ids - for the additional metadata of the new
constraints.

The reason for using names instead of ids, is that some operations - like
store copy - can compact the token id space and change the ids, which
would cause the constraint to have a different - non deterministic -
generated name for an exact same constraint.

In order to do this, this commit applies a refactoring to pass the
TokenNameLookup object instead of expecting all the resolved names
already. It'll then resolve the names as necessary.

We also add an alternative call, without the lookup object, for places
that don't target labels, relationship types nor property keys. These
are, for example, indexes that relies on the
`ANY_TOKEN_NODE_SCHEMA_DESCRIPTOR`.