Fixes for startOffset envelope for EnvelopeChannels
One more allowed value for currentEnvelopeStart - if the buffer
is bigger than segmentSize, the header will also be in the buffer.

Write startOffset envelopes without kernel version - the startOffset
envelope is used when bytes are written directly to the tx log without
parsing the data. It should be possible to not know the kernel version
before doing this appending.