Fix NPE in consistency check ()
This PR fixes NPE in consistency possible when GenericKey is used before it was initialized (i.e. set to some value or read into).
Problem occurs when in visitChildren we encounter internal node with zero keys and only one child. This is still valid but suboptimal tree structure and it can rarely happen. And apparently we have one customer that manages consistently hit it.

To simulate this behaviour in tests I changed SimpleByteArrayLayout to throw exceptions on noninitialized keys in few places.