allow ALL_ENTRIES query to fulltext index
The purpose the the query is to allow a scan over the index.
In most cases, this is more useful for debugging and testing.

It's odd that fulltext doesn't have this when all others do; it seems
like an oversight. Found whilst specifically trying to dump the contents
of a fulltext index in an adhoc test.

Whilst here, also reorder index concepts in the capability test file to
be consistent elsewhere.