remove unused multiple FulltextSearchPredicate functionality.
You can only query fulltext via the procedure which explicitly only uses
a single predicate. We do not use this functionality.

After some discussion, we are removing it, simplifying codepaths.

It is unclear at this time if we need the query to be a boolean query
with 'should occur' clause, or if the raw query is acceptable.

It is possible that with the addition of ALL_ENTRIES in the
previous commit, we might be able to refactor and inherit from
AbstractLuceneIndexReader like the rest of the Lucene based indexes.