Introduce Bolt 5.7 with GQL Compliant Errors ()
To comply with GQL, the failures happen on the database should be exposed along with the `GQLSTATUS` and related information, such as the status description.

Therefore, the `FAILURE`message will need to be updated. This changes should not impact previous version of the protocol, which should work as usual in format and in content. 

## Bolt 5.7

On this version of the protocol, the failure message is re-defined like:

```
FAILURE {
	gql_status::String
	description::String
	message::String
	neo4j_code::String

	// this field is optional and can be ommited in case of default
	diagnostic_record::Dictionary(
		OPERATION::String
		OPERATION_CODE::String
		CURRENT_SCHEMA::String
		_classification::String
		_status_parameters::Dictionary
	)
	
	// This field is optional and will only be included if there is a cause
	cause::Dictionary(
		gql_status::String
		description::String
		message::String

		// this field is optional and can be ommited in case of default
		diagnostic_record::Dictionary(
			OPERATION::String
			OPERATION_CODE::String
			CURRENT_SCHEMA::String
			_classification::String
			_status_parameters::Dictionary
		)
		
		// This field is optional and will only be included if there is a cause
		cause::Dictionary(...)
	)
)
```

Fields on the `diagnostic_record` or the entire object should be omitted if filled with default values. This defaults are versioned along with the protocol version, so newer protocol version can define different defaults. This also applies for `diagnostic_record` on causes.

- Bolt 5.7 Default Diagnostic Record:

```json
{
	"OPERATION": "",
	"OPERATION_CODE": "0",
	"CURRENT_SCHEMA": "/"
}
```

## Design and Implementation notes

The main changes made for implementing this feature were done in the classes `Error`, `FailureMessage` and `FailureMessageEncoder`. 

* `FailureMessageEncoder`  is responsible to encode the message in the respective protocol version. The previous implementation was moved to `FailureMessageEncoderV40` and this one is used for protocol versions older than `5.7`.
* `FailureMessage` holds all the values needed for all serialisable formats of the message. 
* `Error` holds the logic of aggregating information about the exception and then transform to `FailureMessage`, including transversing the GQL causes.

The separation was done on this way to avoid spreading the logic of creating the message over different places on the code base. Ideally, Error should not expose a factory method which doesn't receive a Throwable, but this work will be done in future PR along with the implementation of the GQLSTATUS related to those existing usage of the problematic factory.

### Fabric Support

This feature must be support when using remote databases, so the mapping between driver and fabric/server exceptions need to be update for mapping GQL information coming from the driver.