Introduce Bolt 5.8 ()
This version introduces a `ssr.enabled` as connection hint on the return of `HELLO` and the `advertised_address` in the metadata of `LOGON` response. 

### LOGON

The `advertised_address` was added to the success response. This is the same address of the server in the routing table. It is important to notice that connectors that don't support routing don't return this information. 

```
SUCCESS {
   advertised_address:String?

   // the other properties was not changed
   ...
}
```
### HELLO

The `ssr.enabled` was added to the possible connections hints. This hint indicates if server side routing is enabled in the server. It should be considered false if absent.