Add internal configuration for Max/Min Bolt Versions ()
The goal of this configuration is to provide a method for restricting protocol versions.
This configuration is set as internal usage. 

The configurations entries are `internal.dbms.bolt.min_protocol_version` and `internal.dbms.bolt.max_protocol_version`. They can be configured one independent of the other. Their format are `<major>.<minor>` and the range is inclusive.