Bolt: echo back resolved home DB on TX begin ()
# Bolt 5.8: `db` Field Echoing the User's Resolved Home Database
When receiving a `RUN` or `BEGIN` message in the `READY` state (auto-commit/explicit transaction respectively) with the chosen database being `""` or `null`, the transaction will be execute on the user's home database (or the system's default database).

However, up until now, the client would not be told what this resolved database is. To still be able to make sure that a single driver session runs against the same database for its lifetime, drivers used to manually resolve the database (if not provided by the user) via a `ROUTE` request. This change enables drivers to save this additional `ROUTE` round-trip whose sole purpose is the resolution of the home/default database.

The field is only present if the client did not explicitly provide a database to start the transaction for.

## Example Bolt Exchanges
(assuming in `READY` state for each example)
```
C: BEGIN {}
S: SUCCESS {"db": "yourHomeDb"}
```

```
C: RUN "RETURN 1 AS n" {} {}
S: SUCCESS {"t_first": 3, "fields": ["n"], "db": "yourHomeDb"}
```

```
C: RUN "RETURN 1 AS n" {} {"db": "myChosenDb"}
S: SUCCESS {"t_first": 3, "fields": ["n"]}
```

Signed-off-by: Antonio Barcelos <antonio.barcelos@neo4j.com>
Co-authored-by: Antonio Barcélos <antonio.barcelos@neo4j.com>